Fix action compatibility with npm
When this GitHub Action is used to `yarn install` dependencies, all
works as expected. When the same is performed with `npm install`, 401 is
returned for packages that come from Nexus, internal package registry.

Adding leading slashes will fix this issue by making sure that the
appropriate syntax is used to make access to Nexus npm compatible.